livekit-agent/requirements.txt
livekit-agents[openai,silero,turn-detector,deepgram,cartesia]
livekit-plugins-noise-cancellation
python-dotenv
